+++ include/TinyEngine.hpp
@@ -214,7 +214,7 @@ class TinyEngine {
 
 
 
-	std::vector<uint16_t> indices = {
+	std::vector<uint32_t> indices = {
 		16, 17, 18, 16, 18, 19, // Front
 4,   5,  6,  4,  6,  7, // Back
 0,   1,  2,  0,  2,  3, // Top
